# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Averis Shipping Verification",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/session/actor": {
      "post": {
        "summary": "Switch",
        "operationId": "switch_api_session_actor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cases": {
      "get": {
        "summary": "Cases",
        "operationId": "cases_api_cases_get",
        "parameters": [
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "View"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Category"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Assignee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}": {
      "get": {
        "summary": "Case",
        "operationId": "case_api_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/actions": {
      "post": {
        "summary": "Action",
        "operationId": "action_api_cases__case_id__actions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}/content": {
      "get": {
        "summary": "Content",
        "operationId": "content_api_documents__doc_id__content_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}/preview": {
      "get": {
        "summary": "Preview",
        "operationId": "preview_api_documents__doc_id__preview_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports": {
      "post": {
        "summary": "Import Email",
        "operationId": "import_email_api_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_email_api_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/revisions": {
      "post": {
        "summary": "Revision Upload",
        "operationId": "revision_upload_api_cases__case_id__revisions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_revision_upload_api_cases__case_id__revisions_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "kind": {
            "type": "string",
            "enum": [
              "category",
              "pair",
              "correct",
              "assign",
              "workflow",
              "retry",
              "revision"
            ],
            "title": "Kind"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "BL_COMPARISON",
                  "SI_REQUEST",
                  "INVOICE_QUERY",
                  "GENERAL",
                  "SPAM"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "si_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Si Id"
          },
          "bl_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bl Id"
          },
          "field": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "shipper",
                  "consignee",
                  "notify_party",
                  "port_of_loading",
                  "port_of_discharge",
                  "container_count",
                  "gross_weight_kg"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "maxLength": 256
            },
            "type": "array",
            "maxItems": 100,
            "title": "Evidence Ids"
          },
          "transcription": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16000
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcription"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          },
          "reason": {
            "type": "string",
            "maxLength": 4000,
            "title": "Reason",
            "default": ""
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "workflow": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "open",
                  "waiting",
                  "completed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "kind"
        ],
        "title": "Action"
      },
      "ActorInput": {
        "properties": {
          "actor": {
            "type": "string",
            "title": "Actor"
          }
        },
        "type": "object",
        "required": [
          "actor"
        ],
        "title": "ActorInput"
      },
...
      "Body_import_email_api_imports_post": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "email",
          "files"
        ],
        "title": "Body_import_email_api_imports_post"
      },
      "Body_revision_upload_api_cases__case_id__revisions_post": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "expected_revision",
          "document_id",
          "reason",
          "file"
        ],
        "title": "Body_revision_upload_api_cases__case_id__revisions_post"
      },
      "CasePageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "CasePageResponse"
      },
      "CaseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "received_at": {
            "type": "string",
            "title": "Received At"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "input_revision": {
            "type": "integer",
            "title": "Input Revision"
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Classification"
              },
              {
                "type": "null"
              }
            ]
          },
          "processing": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Processing"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "processing_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Run Id"
          },
          "processing_attempts": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Processing Attempts",
            "default": 0
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "workflow": {
            "type": "string",
            "enum": [
              "open",
              "waiting",
              "completed"
            ],
            "title": "Workflow"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentView"
            },
            "type": "array",
            "title": "Attachments"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Report"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "type": "array",
            "title": "History"
          },
          "summary": {
            "$ref": "#/components/schemas/CaseSummary",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "subject",
          "sender",
          "body",
          "received_at",
          "revision",
          "input_revision",
          "processing",
          "stage",
          "workflow",
          "assignee",
          "review_reasons",
          "attachments",
          "history",
          "summary"
        ],
        "title": "CaseResponse"
      },
...
      "Classification": {
        "properties": {
          "suggested": {
            "type": "string",
            "enum": [
              "BL_COMPARISON",
              "SI_REQUEST",
              "INVOICE_QUERY",
              "GENERAL",
              "SPAM"
            ],
            "title": "Suggested"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Probabilities"
          },
          "accepted": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "BL_COMPARISON",
                  "SI_REQUEST",
                  "INVOICE_QUERY",
                  "GENERAL",
                  "SPAM"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted"
          },
          "source": {
            "type": "string",
            "enum": [
              "model",
              "human",
              "fixture"
            ],
            "title": "Source",
            "default": "model"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "default": "classification-v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suggested",
          "confidence",
          "probabilities",
          "model"
        ],
        "title": "Classification"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionView": {
        "properties": {
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "live_enabled": {
            "type": "boolean",
            "title": "Live Enabled"
          },
          "reviewers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewers",
            "default": [
              "John Tan",
              "Aisha Rahman",
              "Mei Lin"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actor",
          "csrf_token",
          "expires_at",
          "live_enabled"
        ],
        "title": "SessionView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}